requirements.txt
streamlit
gspread
google-auth
oauth2client
pandas
numpy
PyYAML